no hacer nada con newrelic si no hay licencia

+++ .openshift/action_hooks/newrelic/newrelic.sh
@@ -1,9 +1,10 @@
 #!/bin/bash
 
+if [ "${NEW_RELIC_LICENSE_KEY}" != "" ]; then
 CONFIG_FILE=./newrelic.yml
 
 rm -f ${CONFIG_FILE}
 erb ${CONFIG_FILE}.erb > ${CONFIG_FILE}
 
 export CATALINA_OPTS="${CATALINA_OPTS} -javaagent:${OPENSHIFT_REPO_DIR}/.openshift/action_hooks/newrelic/newrelic.jar"
-
+fi
